Take three
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@15301 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/request/NPCPurchaseRequest.java
@@ -188,7 +188,7 @@ public class NPCPurchaseRequest
 		long factor = 100;
 		if ( NPCPurchaseRequest.usingTrousers() ) factor -= 5;
 		if ( KoLCharacter.hasSkill( "Five Finger Discount" ) ) factor -= 5;
-		return ( price * factor ) / 100;
+		return (int) ( ( price * factor ) / 100 );
 	}
 
 	private static boolean usingTrousers()
